#6 refactor: 패키지 구조 변경 및 GlobalExceptionHandler 도입
