+++ src/main/java/com/example/bankingserver/core/account/entity/Account.java
@@ -1,0 +1,38 @@
+package com.example.bankingserver.core.account.entity;
+
+import com.example.bankingserver.core.user.entity.Users;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+public class Account {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Users user;
+
+    private int balance;
+
+    @Builder
+    public Account(Users user, int balance) {
+        this.user = user;
+        this.balance = balance;
+    }
+
+    public void add(int money) {
+        balance += money;
+    }
+
+    public void subtract(int money) {
+        balance -= money;
+    }
+
+}
