+++ src/main/java/com/example/bankingserver/core/account/repository/AccountRepository.java
@@ -1,0 +1,18 @@
+package com.example.bankingserver.core.account.repository;
+
+import com.example.bankingserver.core.account.entity.Account;
+import jakarta.persistence.LockModeType;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Lock;
+import org.springframework.data.jpa.repository.Query;
+
+import java.util.Optional;
+
+public interface AccountRepository extends JpaRepository<Account, Long> {
+    Optional<Account> findByUserId(Long userId);
+
+    @Lock(value = LockModeType.PESSIMISTIC_WRITE)
+    @Query("select a from Account a where a.id = :accountId")
+    Optional<Account> findByIdWithPessimisticLock(Long accountId);
+
+}
