+++ src/main/java/com/example/bankingserver/core/friendship/repository/FriendshipRepository.java
@@ -1,0 +1,11 @@
+package com.example.bankingserver.core.friendship.repository;
+
+import com.example.bankingserver.core.friendship.entity.Friendship;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface FriendshipRepository extends JpaRepository<Friendship, Long> {
+
+    Optional<Friendship> findByUserIdAndFriendId(Long userId, Long friendId);
+}
