+++ src/main/java/com/example/bankingserver/core/transaction/repository/TransactionRepository.java
@@ -1,0 +1,7 @@
+package com.example.bankingserver.core.transaction.repository;
+
+import com.example.bankingserver.core.transaction.entity.Transaction;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+}
