+++ src/main/java/com/example/bankingserver/core/user/dto/request/UserJoinRequestDto.java
@@ -1,0 +1,28 @@
+package com.example.bankingserver.core.user.dto.request;
+
+import com.example.bankingserver.core.user.entity.Users;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class UserJoinRequestDto {
+
+    private String username;
+    private String password;
+
+    @Builder
+    public UserJoinRequestDto(String username, String password) {
+        this.username = username;
+        this.password = password;
+    }
+
+    public Users toEntity() {
+        return Users.builder()
+                .username(username)
+                .password(password)
+                .build();
+    }
+}
