+++ src/main/java/com/example/bankingserver/core/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.example.bankingserver.core.user.repository;
+
+import com.example.bankingserver.core.user.entity.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<Users, Long> {
+    Optional<Users> findByUsername(String username);
+}
