--- src/main/java/com/example/bankingserver/domain/AccountRepository.java
@@ -1,17 +1,0 @@
-package com.example.bankingserver.domain;
-
-import jakarta.persistence.LockModeType;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.Optional;
-
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    Optional<Account> findByUserId(Long userId);
-
-    @Lock(value = LockModeType.PESSIMISTIC_WRITE)
-    @Query("select a from Account a where a.id = :accountId")
-    Optional<Account> findByIdWithPessimisticLock(Long accountId);
-
-}
