--- src/main/java/com/example/bankingserver/domain/TransactionRepository.java
@@ -1,6 +1,0 @@
-package com.example.bankingserver.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
-}
