--- src/main/java/com/example/bankingserver/domain/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.bankingserver.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<Users, Long> {
-    Optional<Users> findByUsername(String username);
-}
