--- src/main/java/com/example/bankingserver/exception/AccountNotFoundException.java
@@ -1,12 +1,0 @@
-package com.example.bankingserver.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class AccountNotFoundException extends RuntimeException {
-
-    public AccountNotFoundException(String message) {
-        super(message);
-    }
-}
