--- src/main/java/com/example/bankingserver/exception/BadRequestException.java
@@ -1,12 +1,0 @@
-package com.example.bankingserver.exception;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.BAD_REQUEST)
-public class BadRequestException extends RuntimeException {
-
-    public BadRequestException(String message) {
-        super(message);
-    }
-}
