--- src/main/java/com/example/bankingserver/exception/global/BasicExceptionType.java
@@ -1,13 +1,0 @@
-package com.example.bankingserver.exception.global;
-
-import org.springframework.http.HttpStatus;
-
-public interface BasicExceptionType {
-
-    String getErrorCode();
-
-    String getMessage();
-
-    HttpStatus getHttpStatus();
-
-}
