--- src/main/java/com/example/bankingserver/exception/global/BusinessLogicException.java
@@ -1,15 +1,0 @@
-package com.example.bankingserver.exception.global;
-
-import lombok.Getter;
-
-@Getter
-public class BusinessLogicException extends RuntimeException {
-
-    private final BasicExceptionType basicExceptionType;
-
-    public BusinessLogicException(BasicExceptionType basicExceptionType) {
-
-        super(basicExceptionType.getMessage());
-        this.basicExceptionType = basicExceptionType;
-    }
-}
