+++ src/main/java/com/example/bankingserver/global/dto/RsData.java
@@ -1,0 +1,24 @@
+package com.example.bankingserver.global.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class RsData<T> {
+    private String resultCode;
+    private String msg;
+    private T data;
+
+    public static <T> RsData<T> of(String resultCode, String msg) {
+        return new RsData<>(resultCode, msg, null);
+    }
+
+    public static <T> RsData<T> of(String resultCode, String msg, T data) {
+        return new RsData<>(resultCode, msg, data);
+    }
+
+}
+
