+++ src/main/java/com/example/bankingserver/global/exception/BasicExceptionType.java
@@ -1,0 +1,13 @@
+package com.example.bankingserver.global.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface BasicExceptionType {
+
+    String getErrorCode();
+
+    String getMessage();
+
+    HttpStatus getHttpStatus();
+
+}
