+++ src/main/java/com/example/bankingserver/global/exception/BusinessLogicException.java
@@ -1,0 +1,15 @@
+package com.example.bankingserver.global.exception;
+
+import lombok.Getter;
+
+@Getter
+public class BusinessLogicException extends RuntimeException {
+
+    private final BasicExceptionType basicExceptionType;
+
+    public BusinessLogicException(BasicExceptionType basicExceptionType) {
+
+        super(basicExceptionType.getMessage());
+        this.basicExceptionType = basicExceptionType;
+    }
+}
